requirements.txt
astroid==2.2.5
beautifulsoup4==4.7.1
boto3==1.9.188
botocore==1.12.188
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==2.1
django-allauth==0.39.1
django-debug-toolbar==2.0
django-elasticsearch-dsl==0.5.1
django-extensions==2.1.9
django-filter==2.1.0
django-rest-swagger==2.1.2
django-storages==1.7.1
djangorestframework==3.9.4
docutils==0.14
drf-yasg==1.16.0
elasticsearch==6.4.0
elasticsearch-dsl==6.1.0
et-xmlfile==1.0.1
idna==2.8
inflection==0.3.1
ipaddress==1.0.22
isort==4.3.16
itypes==1.1.0
jdcal==1.4
Jinja2==2.10.1
jmespath==0.9.4
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector-python==8.0.15
oauthlib==3.0.1
openapi-codec==1.3.2
openpyxl==2.6.1
Pillow==6.1.0
pipenv==2018.11.26
powerline-status==2.7
protobuf==3.7.1
psycopg2-binary==2.8.3
pygraphviz==1.5
pylint==2.3.1
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
ruamel.yaml==0.15.98
s3transfer==0.2.1
simplejson==3.16.0
six==1.12.0
soupsieve==1.9.1
sqlparse==0.3.0
typed-ast==1.3.1
uritemplate==3.0.0
urllib3==1.25.3
virtualenv==16.5.0
virtualenv-clone==0.5.3
wrapt==1.11.1
